Refactor sendTransactionLinkredeemed .
