+++ backend/src/mailer/sendTransactionReceivedEmail.test.ts
@@ -19,7 +19,6 @@ describe('sendTransactionReceivedEmail', () => {
       email: 'peter@lustig.de',
       senderEmail: 'bibi@bloxberg.de',
       amount: new Decimal(42.0),
-      memo: 'Vielen herzlichen Dank für den neuen Hexenbesen!',
       overviewURL: 'http://localhost/overview',
     })
   })
@@ -33,7 +32,6 @@ describe('sendTransactionReceivedEmail', () => {
         expect.stringContaining('42,00 GDD') &&
         expect.stringContaining('Bibi Bloxberg') &&
         expect.stringContaining('(bibi@bloxberg.de)') &&
-        expect.stringContaining('Vielen herzlichen Dank für den neuen Hexenbesen!') &&
         expect.stringContaining('http://localhost/overview'),
     })
   })
