+++ backend/src/mailer/sendTransactionReceivedEmail.ts
@@ -1,7 +1,7 @@
 import { backendLogger as logger } from '@/server/logger'
 import Decimal from 'decimal.js-light'
 import { sendEMail } from './sendEMail'
-import { transactionLinkRedeemed, transactionReceived } from './text/transactionReceived'
+import { transactionReceived } from './text/transactionReceived'
 
 export const sendTransactionReceivedEmail = (data: {
   senderFirstName: string
@@ -11,7 +11,6 @@ export const sendTransactionReceivedEmail = (data: {
   email: string
   senderEmail: string
   amount: Decimal
-  memo: string
   overviewURL: string
 }): Promise<boolean> => {
   logger.info(
@@ -24,29 +23,5 @@ export const sendTransactionReceivedEmail = (data: {
     to: `${data.recipientFirstName} ${data.recipientLastName} <${data.email}>`,
     subject: transactionReceived.de.subject,
     text: transactionReceived.de.text(data),
-  })
-}
-
-export const sendTransactionLinkRedeemedEmail = (data: {
-  email: string
-  senderFirstName: string
-  senderLastName: string
-  recipientFirstName: string
-  recipientLastName: string
-  senderEmail: string
-  amount: Decimal
-  memo: string
-  overviewURL: string
-}): Promise<boolean> => {
-  logger.info(
-    `sendEmail(): to=${data.recipientFirstName} ${data.recipientLastName},
-      <${data.email}>,
-      subject=${transactionLinkRedeemed.de.subject},
-      text=${transactionLinkRedeemed.de.text(data)}`,
-  )
-  return sendEMail({
-    to: `${data.recipientFirstName} ${data.recipientLastName} <${data.email}>`,
-    subject: transactionLinkRedeemed.de.subject,
-    text: transactionLinkRedeemed.de.text(data),
   })
 }
