+++ backend/src/mailer/text/transactionReceived.ts
@@ -11,7 +11,6 @@ export const transactionReceived = {
       email: string
       senderEmail: string
       amount: Decimal
-      memo: string
       overviewURL: string
     }): string =>
       `Hallo ${data.recipientFirstName} ${data.recipientLastName}
@@ -19,38 +18,6 @@ export const transactionReceived = {
 Du hast soeben ${data.amount.toFixed(2).replace('.', ',')} GDD von ${data.senderFirstName} ${
         data.senderLastName
       } (${data.senderEmail}) erhalten.
-
-Details zur Transaktion findest du in deinem Gradido-Konto: ${data.overviewURL}
-
-Bitte antworte nicht auf diese E-Mail!
-
-Mit freundlichen Grüßen,
-dein Gradido-Team`,
-  },
-}
-
-export const transactionLinkRedeemed = {
-  de: {
-    subject: 'Gradido link eingelösst',
-    text: (data: {
-      email: string
-      senderFirstName: string
-      senderLastName: string
-      recipientFirstName: string
-      recipientLastName: string
-      senderEmail: string
-      amount: Decimal
-      memo: string
-      overviewURL: string
-    }): string =>
-      `Hallo ${data.recipientFirstName} ${data.recipientLastName}
-
-${data.senderFirstName} ${data.senderLastName} (${
-        data.senderEmail
-      }) hat soeben deinen Link eingelöst.
-
-Betrag: ${data.amount.toFixed(2).replace('.', ',')} GDD,
-Memo: ${data.memo}
 
 Details zur Transaktion findest du in deinem Gradido-Konto: ${data.overviewURL}
